Refine redis queue pub/sub
