+++ trapdoor/routes.py
@@ -6,11 +6,8 @@ HANDLERS = [
     (r"/resolve/?", handlers.Resolve),
     (r"/resolve_all/?", handlers.ResolveAll),
 
-    #messageWall
-    (r"/wall", handlers.Wall),
-
     #websocket
-    (r"/socket", utils.ChatSocketHandler),
+    (r"/ChatSocket", handlers.ChatSocketHandler),
 
     # API
     (r"/api/varbinds/(?P<notification_id>\d+)", handlers.ApiVarBinds),
